refactor: replace magic numbers in CPU display with named constants
- CPU_X=90, CPU_WARN_PCT=80 → display_layout.h
- CPU_AVERAGE_FRAMES = 1000/DISPLAY_UPDATE_MS → constants.h
- display_last_ms declared static to match other loop-persistent vars
- GetCpuUsage() uses explicit static_cast<float> on volatile read

+++ src/audio/audio_engine.h
@@ -32,7 +32,7 @@ public:
     void SetHold(bool hold);
     // Intentionally unsynchronised: a 4-byte aligned volatile float read is
     // atomic on Cortex-M7; stale-by-one-block is acceptable for a 30 Hz display.
-    static float GetCpuUsage() { return cpu_usage_; }
+    static float GetCpuUsage() { return static_cast<float>(cpu_usage_); }
 
     static void AudioCallback(daisy::AudioHandle::InputBuffer  in,
                               daisy::AudioHandle::OutputBuffer out,
